Images fix readme

+++ server/app/post.py
@@ -76,6 +76,9 @@ def create():
             # when saving, we need absolute path
             img_abs_path = os.path.join(IMAGE_UPLOAD_DIRECTORY, filename)
 
+            if not os.path.exists(IMAGE_UPLOAD_DIRECTORY):
+                os.mkdir(IMAGE_UPLOAD_DIRECTORY)
+
             # when displaying in html, we need relative path
             img_rel_path = os.path.join(IMAGE_UPLOAD_REL_DIRECTORY, filename)
 
@@ -108,7 +111,7 @@ def edit(post_id):
 
     if request.method == "POST":
         post_text = request.form["posttext"]
-        post_has_image = "image" in request.files
+        post_has_image = "image" in request.files and request.files["image"].filename != ""
 
         if post_text == "":
             error = "You cannot post nothing!"
@@ -139,13 +142,13 @@ def edit(post_id):
 
         # upload the new image and delete the old one
         if post_has_image:
-
             # Check if the directory already exists
             if not os.path.exists(IMAGE_UPLOAD_DIRECTORY):
                 # Create the directory, also create intermediate directories if necessary
                 os.makedirs(IMAGE_UPLOAD_DIRECTORY)
 
-            os.remove(os.path.join(REAL_PATH, post.image))
+            if post.image:
+                os.remove(os.path.join(REAL_PATH, post.image))
             file = request.files["image"]
             if file.filename == "" or not is_file_allowed(file.filename):
                 flash("That file is not valid")
@@ -159,12 +162,8 @@ def edit(post_id):
             img_rel_path = os.path.join(IMAGE_UPLOAD_REL_DIRECTORY, filename)
 
             file.save(img_abs_path)
-            db.session.add(Post(user_id=user.id, username=user.username,
-                                content=post_text, date_time=datetime.now(), 
-                                image=img_rel_path))
-        else:
-            db.session.add(Post(user_id=user.id, username=user.username,
-                                content=post_text, date_time=datetime.now()))
+            post.image = img_rel_path
+        post.content = post_text
         db.session.commit()
         return redirect(url_for("feed.feed"))
     return render_template("createpost.html", text=post.content)
